LPS-131825 Refactor fieldSet add action
